Move FileTreeRepository to the internal package
Until we want to expose these interfaces to users, they should be in the
internal package.
