+++ io/src/main/scala/sbt/internal/io/SourceModificationWatch.scala
@@ -17,7 +17,7 @@ import java.util.concurrent.ConcurrentHashMap
 import java.util.concurrent.atomic.AtomicBoolean
 
 import sbt.internal.io.FileEvent.{ Deletion, Update }
-import sbt.io.FileTreeView.AllPass
+import FileTreeView.AllPass
 import sbt.io._
 import sbt.io.syntax._
