+++ io/src/main/scala/sbt/io/Path.scala
@@ -25,8 +25,8 @@ import java.nio.file.{
 
 import com.swoval.files.FileTreeViews
 import com.swoval.functional.Filter
-import sbt.internal.io.SimpleFileAttributes
-import sbt.io.FileTreeView.AllPass
+import sbt.internal.io.{ FileTreeView, SimpleFileAttributes }
+import sbt.internal.io.FileTreeView.AllPass
 
 import scala.collection.JavaConverters._
 import scala.collection.mutable
